Add files via upload
Creating a program that deals with a mass shooting incident, especially when the data is highly sensitive and has gone missing, is an exceedingly complex task. However, it's possible to conceptualize a basic Python program that simulates a simple incident data management and recovery system.

This program will involve two classes, Incident and IncidentDatabase. The Incident class will represent individual incidents, and the IncidentDatabase class will manage a collection of incidents, including adding new incidents and attempting to recover missing incidents.